pyproject.toml
[project]
name = "circus"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "BeaverNotACat", email = "beavernotacat@gmail.com" }
]
dependencies = [
    "dependency-injector>=4.43.0",
    "pillow>=11.0.0",
    "pydantic-settings>=2.6.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/circus"]
